// app/frontend/src/context/api/models/TodoPriority.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HiveSync API 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const TodoPriority = {
    Low: 'Low',
    Medium: 'Medium',
    High: 'High'
} as const;
export type TodoPriority = typeof TodoPriority[keyof typeof TodoPriority];


export function instanceOfTodoPriority(value: any): boolean {
    for (const key in TodoPriority) {
        if (Object.prototype.hasOwnProperty.call(TodoPriority, key)) {
            if (TodoPriority[key as keyof typeof TodoPriority] === value) {
                return true;
            }
        }
    }
    return false;
}

export function TodoPriorityFromJSON(json: any): TodoPriority {
    return TodoPriorityFromJSONTyped(json, false);
}

export function TodoPriorityFromJSONTyped(json: any, ignoreDiscriminator: boolean): TodoPriority {
    return json as TodoPriority;
}

export function TodoPriorityToJSON(value?: TodoPriority | null): any {
    return value as any;
}

export function TodoPriorityToJSONTyped(value: any, ignoreDiscriminator: boolean): TodoPriority {
    return value as TodoPriority;
}